Add files via upload
the website built with this project has the goal of enabling sw developers to discuss and report problems. So it's kind of a forum, as the JSP class says.
The technologies used for this project are multiple, from JSPs and taglibs to handle the languages, to CSS for the visual impact and UX.
Last but not least the use of MySQL is pretty much everywhere, instead of using only cookies and session attributes.
It has been developed by my group for the WebCoding Exam.